+++ Frends.Files.Delete/Frends.Files.Delete/Delete.cs
@@ -14,6 +14,7 @@ using System.Reflection;
 using System.Security.Principal;
 using System.Threading;
 using System.Text.RegularExpressions;
+using Frends.Files.Delete.Helpers;
 
 namespace Frends.Files.Delete;
 
@@ -92,12 +93,12 @@ public class Files
 
     private static List<FileItem> ExecuteDelete(Input input, CancellationToken cancellationToken)
     {
-        var results = FindMatchingFiles(input.Directory, input.Pattern);
+        var results = FilesHandler.FindMatchingFiles(input.Directory, input.Pattern);
 
         var fileResults = new List<FileItem>();
         try
         {
-            foreach (var path in results.Files.Select(match => Path.Combine(input.Directory, match.Path)))
+            foreach (var path in results.Select(match => Path.Combine(input.Directory, match)))
             {
                 cancellationToken.ThrowIfCancellationRequested();
 
@@ -121,33 +122,6 @@ public class Files
             throw new ArgumentException($@"UserName field must be of format domain\username was: {username}");
 
         return new Tuple<string, string>(domainAndUserName[0], domainAndUserName[1]);
-    }
-
-    internal static PatternMatchingResult FindMatchingFiles(string directoryPath, string pattern)
-    {
-        // Check the user can access the folder.
-        // This will return false if the path does not exist or you do not have read permissions.
-        if (!Directory.Exists(directoryPath))
-            throw new DirectoryNotFoundException($"Directory does not exist or you do not have read access. Tried to access directory '{directoryPath}'");
-
-        if (pattern.StartsWith("<regex>"))
-        {
-            string regexPattern = pattern.Substring(7);
-
-            var matchingFiles = Directory.GetFiles(directoryPath, "*", SearchOption.AllDirectories)
-                .Where(file => Regex.IsMatch(Path.GetFileName(file), regexPattern))
-                .Select(file => new FilePatternMatch(Path.GetFileName(file), Path.GetFileName(file)))
-                .ToList();
-
-            return new PatternMatchingResult(matchingFiles);
-        }
-        else
-        {
-            var matcher = new Matcher();
-            matcher.AddInclude(pattern);
-            var results = matcher.Execute(new DirectoryInfoWrapper(new DirectoryInfo(directoryPath)));
-            return results;
-        }
     }
 
     private static void OnPluginUnloadingRequested(AssemblyLoadContext obj)