+++ Frends.Files.Move/Frends.Files.Move.Tests/RemoteTests.cs
@@ -1,10 +1,12 @@
 ﻿using System;
 using System.IO;
+using System.Runtime.Versioning;
 using System.Security.Principal;
 using Frends.Files.Move.Definitions;
 using NUnit.Framework;
 using System.Threading;
 using System.Threading.Tasks;
+using dotenv.net;
 using SimpleImpersonation;
 
 namespace Frends.Files.Move.Tests;
@@ -15,31 +17,42 @@ internal class RemoteTests
     private static readonly string LocalWorkdir =
         Path.GetFullPath(Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "TestData"));
 
-    private static readonly string SrcUser = Environment.GetEnvironmentVariable("SRC_USER")!;
+    internal required string SrcUser { get; set; }
+    internal required string SrcUserPassword { get; set; }
+    internal required string DstUser { get; set; }
+    internal required string DstUserPassword { get; set; }
+    internal required string AdminUser { get; set; }
+    internal required string AdminUserPassword { get; set; }
+    internal required string RemoteIp { get; set; }
+    internal required string Domain { get; set; }
 
-    private static readonly string
-        SrcUserPassword = Environment.GetEnvironmentVariable("SRC_PASSWORD")!;
+    [OneTimeSetUp]
+    public void Setup()
+    {
+        DotEnv.Load();
+        SrcUser = GetEnvVar("SRC_USER");
+        SrcUserPassword = GetEnvVar("SRC_PASSWORD");
+        DstUser = GetEnvVar("DST_USER");
+        DstUserPassword = GetEnvVar("DST_PASSWORD");
+        AdminUser = GetEnvVar("ADMIN_USER");
+        AdminUserPassword = GetEnvVar("ADMIN_PASSWORD");
+        RemoteIp = GetEnvVar("REMOTE_IP");
+        Domain = GetEnvVar("DOMAIN");
+    }
 
-    private static readonly string
-        DstUser = Environment.GetEnvironmentVariable("DST_USER")!;
-
-    private static readonly string
-        DstUserPassword = Environment.GetEnvironmentVariable("DST_PASSWORD")!;
-
-    private static readonly string
-        AdminUser = Environment.GetEnvironmentVariable("ADMIN_USER")!;
-
-    private static readonly string
-        AdminUserPassword = Environment.GetEnvironmentVariable("ADMIN_PASSWORD")!;
+    private static string GetEnvVar(string name) =>
+        Environment.GetEnvironmentVariable(name) ??
+        throw new InvalidOperationException($"Missing required env var: {name}");
 
     [Test]
+    [SupportedOSPlatform("windows")]
     public async Task MoveFileFromLocalToRemoteWithImpersonation()
     {
         var input = new Input
         {
             SourceDirectory = LocalWorkdir,
             Pattern = "*",
-            TargetDirectory = @"\\20.67.234.98\Shared\dst",
+            TargetDirectory = $@"{RemoteIp}\Shared\dst",
         };
         var options = new Options
         {
@@ -48,7 +61,7 @@ internal class RemoteTests
         var connection = new Connection
         {
             TargetIsRemote = true,
-            TargetUserName = DstUser,
+            TargetUserName = $@"{Domain}\{DstUser}",
             TargetPassword = DstUserPassword,
         };
         PrepareSourceAndTarget(input, connection);
@@ -58,11 +71,12 @@ internal class RemoteTests
     }
 
     [Test]
+    [SupportedOSPlatform("windows")]
     public async Task MoveFileFromRemoteToLocalWithImpersonation()
     {
         var input = new Input
         {
-            SourceDirectory = @"\\20.67.234.98\Shared\src",
+            SourceDirectory = $@"{RemoteIp}\Shared\src",
             Pattern = "*",
             TargetDirectory = LocalWorkdir,
         };
@@ -73,7 +87,7 @@ internal class RemoteTests
         var connection = new Connection
         {
             SourceIsRemote = true,
-            SourceUserName = SrcUser,
+            SourceUserName = $@"{Domain}\{SrcUser}",
             SourcePassword = SrcUserPassword,
         };
         PrepareSourceAndTarget(input, connection);
@@ -83,13 +97,14 @@ internal class RemoteTests
     }
 
     [Test]
+    [SupportedOSPlatform("windows")]
     public async Task MoveFileFromRemoteToRemoteWithImpersonation()
     {
         var input = new Input
         {
-            SourceDirectory = @"\\20.67.234.98\Shared\src",
+            SourceDirectory = $@"{RemoteIp}\Shared\src",
             Pattern = "*",
-            TargetDirectory = @"\\20.67.234.98\Shared\dst",
+            TargetDirectory = $@"{RemoteIp}\Shared\dst",
         };
         var options = new Options
         {
@@ -98,10 +113,10 @@ internal class RemoteTests
         var connection = new Connection
         {
             SourceIsRemote = true,
-            SourceUserName = SrcUser,
+            SourceUserName = $@"{Domain}\{SrcUser}",
             SourcePassword = SrcUserPassword,
             TargetIsRemote = true,
-            TargetUserName = DstUser,
+            TargetUserName = $@"{Domain}\{DstUser}",
             TargetPassword = DstUserPassword,
         };
         PrepareSourceAndTarget(input, connection);
@@ -129,9 +144,10 @@ internal class RemoteTests
             File.Delete(targetFilePath);
     }
 
-    private static void PrepareSourceAndTarget(Input input, Connection connection)
+    [SupportedOSPlatform("Windows")]
+    private void PrepareSourceAndTarget(Input input, Connection connection)
     {
-        var (domain, user) = GetDomainAndUsername(AdminUser);
+        var (domain, user) = GetDomainAndUsername($@"{Domain}\{AdminUser}");
         var credentials = new UserCredentials(domain, user, AdminUserPassword);
         using var userHandle = credentials.LogonUser(LogonType.NewCredentials);
         if (connection.SourceIsRemote)
@@ -148,6 +164,7 @@ internal class RemoteTests
     private static Tuple<string, string> GetDomainAndUsername(string username)
     {
         var domainAndUserName = username.Split('\\');
+
         return domainAndUserName.Length != 2
             ? throw new ArgumentException($@"UserName field must be of format domain\username was: {username}")
             : new Tuple<string, string>(domainAndUserName[0], domainAndUserName[1]);