+++ Frends.Files.Move/Frends.Files.Move/Utils.cs
@@ -11,7 +11,7 @@ using Microsoft.Extensions.FileSystemGlobbing.Abstractions;
 
 namespace Frends.Files.Move;
 
-internal static class Helpers
+internal static class Utils
 {
     internal static Tuple<string, string> GetDomainAndUsername(string username)
     {
@@ -19,35 +19,6 @@ internal static class Helpers
         return domainAndUserName.Length != 2
             ? throw new ArgumentException($@"UserName field must be of format domain\username was: {username}")
             : new Tuple<string, string>(domainAndUserName[0], domainAndUserName[1]);
-    }
-
-    internal static PatternMatchingResult FindMatchingFiles(string directoryPath, string pattern)
-    {
-        // This will return false if the path does not exist, or you do not have read permissions.
-        if (!Directory.Exists(directoryPath))
-            throw new DirectoryNotFoundException(
-                $"Directory does not exist or you do not have read access. Tried to access directory '{directoryPath}'");
-
-        if (pattern.StartsWith("<regex>"))
-        {
-            var regexPattern = pattern[7..];
-
-            var matchingFiles = Directory.GetFiles(directoryPath, "*", SearchOption.AllDirectories)
-                .Where(file => Regex.IsMatch(Path.GetFileName(file), regexPattern))
-                .Select(file =>
-                {
-                    var relativePath = Path.GetRelativePath(directoryPath, file);
-                    return new FilePatternMatch(relativePath, relativePath);
-                })
-                .ToList();
-
-            return new PatternMatchingResult(matchingFiles);
-        }
-
-        var matcher = new Matcher();
-        matcher.AddInclude(pattern);
-        var results = matcher.Execute(new DirectoryInfoWrapper(new DirectoryInfo(directoryPath)));
-        return results;
     }
 
     internal static async Task CopyFileImpersonated(string sourceFilePath, string targetFilePath, Connection connection,
@@ -61,15 +32,15 @@ internal static class Helpers
         await sourceStream.CopyToAsync(targetStream, 81920, cancellationToken).ConfigureAwait(false);
     }
 
-    internal static Dictionary<string, string> GetFileTransferEntries(IEnumerable<FilePatternMatch> fileMatches,
+    internal static Dictionary<string, string> GetFileTransferEntries(IEnumerable<string> fileMatches,
         string sourceDirectory, string targetDirectory, bool preserveDirectoryStructure)
     {
         return fileMatches
             .ToDictionary(
-                f => Path.Combine(sourceDirectory, f.Path),
+                f => Path.Combine(sourceDirectory, f),
                 f => preserveDirectoryStructure
-                    ? Path.GetFullPath(Path.Combine(targetDirectory, f.Path))
-                    : Path.GetFullPath(Path.Combine(targetDirectory, Path.GetFileName(f.Path))));
+                    ? Path.GetFullPath(Path.Combine(targetDirectory, f))
+                    : Path.GetFullPath(Path.Combine(targetDirectory, Path.GetFileName(f))));
     }
 
     internal static void AssertNoTargetFileConflicts(string[] filePaths)